requirements.txt
google-ads==17.0.0
google-api-core==2.8.2
google-api-python-client==2.52.0
google-auth==2.9.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
google-cloud==0.34.0
google-cloud-bigquery==3.2.0
google-cloud-bigquery-storage==2.13.2
google-cloud-core==2.3.1
google-cloud-secret-manager==2.12.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.3
pybigquery==0.5.0
SQLAlchemy==1.4.39
numpy==1.23.0
pandas==1.4.3
